requirements.txt
neomodel==5.2.0
python-dotenv==1.0.0
astroid==3.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
